# go-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/checkboxes/current": {
            "get": {
                "description": "Get the current bitset",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the current bitset",
                "operationId": "get-current-bitset",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetCurrentResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/checkboxes/update": {
            "post": {
                "description": "Update the bitset at a given position",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update the bitset at a given position",
                "operationId": "update-bitset",
                "parameters": [
                    {
                        "description": "Update Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/checkboxes/wipe": {
            "get": {
                "description": "Wipe the current bitset",
                "produces": [
                    "application/json"
                ],
                "summary": "Wipe the current bitset",
                "operationId": "wipe-bitset",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetCurrentResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.GetCurrentResponse": {
            "type": "object",
            "properties": {
                "bitSet": {
                    "type": "string"
                }
            }
        },
        "models.UpdateRequest": {
            "type": "object",
            "properties": {
                "position": {
                    "type": "integer"
                },
                "value": {
                    "type": "boolean"
                }
            }
        },
        "models.UpdateResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    }
}